// src/types.d.ts
export interface S316GeneratorElectric {
  kwSize: string;
  peaName: string;
  telNumber: string;
  distanceKm: number;
}
// ...
interface UploaddedImage {
  id: number | string;
  file_id: string;
  name: string;
  fileName: string;
  originalName: string;
  size: number;
  fileSize: number;
  url: string;
  file: File;
  uploadDate?: Date;
  isUploading?: boolean;
  isOffline?: boolean;
}
// ...
interface Survey {
  serviceId: string;
  workOrderNo: string;
  workOrderStatusCode: string;
  customerRequestNo: string;
  sapOrderNo: string;
  serviceName: string;
  requestCode: string;
  requestChannel: string;
  sapProcessCreatedDate?: string;
  customerMobileNo: number;
  customerLatitude: number;
  customerLongitude: number;
  customerAddress: string;
  customerName: string;
  requestData: {
    morning_flag: boolean,
    afternoon_flag: boolean
    detail: string
  },
  surveyData: WorkOrderSurveyData,
  note: string,
  images: number[],

  routeCode: string
  customerNumber: string
  meterSerialNumber: string
  recommendedPhase: string
  mainWorkCenterId: string
}

// S332 Solar Air Survey Data
export interface S332SolarAirSurveyData {
  house_facing_direction: string;
  refrigerant_pipe_length_m: number;
  dc_cable_length_m: number;
  ac_cable_length_m: number;
  house_front_media_id: number;
  houseFrontMedia: UploadedFile;
  roof_installation_media_id: number;
  roofInstallationMedia: UploadedFile;
  outdoor_unit_location_media_id: number;
  outdoorUnitLocationMedia: UploadedFile;
  indoor_unit_location_media_id: number;
  indoorUnitLocationMedia: UploadedFile;
  ac_cabling_route_media_id: number;
  acCablingRouteMedia: UploadedFile;
  dc_cabling_route_media_id: number;
  dcCablingRouteMedia: UploadedFile;
  cutout_breaker_media_id: number;
}

// S332 Solar Battery Survey Data
export interface S332SolarBatterySurveyData {
  pv_to_inverter_dist_m?: number;
  inverter_to_mdb_dist_m?: number;
  internet_conn_type?: string;
  internet_conn_remark?: string;
  houseFrontMedia: UploadedFile;
  meterBillMedia: UploadedFile;
  consumerUnitMedia: UploadedFile;
  roofAreaMedia: UploadedFile;
  inverterLocationMedia: UploadedFile;
  batteryLocationMedia: UploadedFile;
}

export interface BusinessTypeSurveyData {
  business_type_id: string;
}

// S316 Survey Data: ขอเช่าเครื่องกำเนิดไฟฟ้า
export interface S316SurveyData {
  kw_size_reservation: Array<{
    id: number;
    pea_name: string;
    distance: number;
    tel_number: string;
  }>;
  requests_serviceS316: {
    rental_startdate: string | Date | undefined;
    rental_enddate: string | Date | undefined;
    start_time: string;
    end_time: string;
  };
}

export interface WorkOrderSurveyData {
  status: string;
  note: string;
  result_note: string;
  survey_by: string;
  appointment_date: string | Date |  undefined;
  images: UploaddedImage[];
  serviceSpecificData: BusinessTypeSurveyData
    | S316SurveyData
    | S332SolarAirSurveyData
    | S332SolarBatterySurveyData
}